add faulty component

+++ src/app/delete-page/delete-page.component.ts
@@ -7,25 +7,35 @@ import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
   templateUrl: './delete-page.component.html',
   styleUrls: ['./delete-page.component.scss']
 })
-export class DeletePageComponent implements OnInit{
-constructor(
-  @Inject(MAT_DIALOG_DATA) public data : any,
-  private dialogRef: MatDialogRef<DeletePageComponent>,
-  private anotherService: AnotherService
-) {}
+export class DeletePageComponent implements OnInit {
+  constructor(
+    @Inject(MAT_DIALOG_DATA) public data: any,
+    private dialogRef: MatDialogRef<DeletePageComponent>,
+    private anotherService: AnotherService
+  ) { }
 
-ngOnInit(): void {
-}
+  ngOnInit(): void {
+  }
   onDismiss() {
     this.dialogRef.close();
   }
 
   onConfirm() {
-    this.dialogRef.close();
-    this.anotherService.deleteImage(this.data.doc.id).subscribe(res => {
-      console.log(res)
-    })
-    this.dialogRef.close();
-    window.location.reload();
+    if (this.data.action == 'image') {
+      this.anotherService.deleteImage(this.data.doc.id).subscribe(res => {
+        console.log(res)
+      })
+      console.log('image')
+      this.dialogRef.close();
+      window.location.reload();
+    }
+    if (this.data.action == 'faulty') {
+      this.anotherService.deleteFaultyImage(this.data.doc.faultyImgId).subscribe(res => {
+        console.log(res)
+      })
+      console.log('faulty')
+      this.dialogRef.close();
+      window.location.reload();
+    }
   }
 }
